add instructor mode; restructure; some other stuff
