+++ src/components/CheckboxInputComponent.tsx
@@ -1,10 +1,9 @@
 import { ReactElement } from "react"
 
-const LABEL_CLASSNAME = 'text-lg mb-2 w-full'
 const CheckboxComponent = ({formName, title, readOnly, value}: Props): ReactElement => {
     return (
         <div className={"flex flex-col mr-5"}>
-            <label htmlFor={formName?.toLowerCase()} className={LABEL_CLASSNAME}>{title}</label>
+            <label htmlFor={formName?.toLowerCase()}>{title}</label>
             <input name={formName} disabled={readOnly} defaultChecked={value ? true : false} type="checkbox" className={'rounded-sm border-1 text-xl checked:accent-ezblue w-8 h-8 transition-colors'}/>
         </div>
     )
@@ -16,7 +15,6 @@ type Props = {
     formName?: string
     readOnly?: boolean
     value?: boolean
-    defaultValue?: boolean
 }
 
 export default CheckboxComponent
