+++ src/layouts/Layout.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react"
 import { Outlet } from "react-router-dom"
-import Navbar from "./Navbar"
-import Footer from "./Footer"
+import Navbar from "../components/Navbar"
+import Footer from "../views/Footer"
 
 const Layout = (): ReactElement => {
     return (
