+++ src/views/instructor/InstructorPanel.tsx
@@ -1,0 +1,11 @@
+import React, { ReactNode } from 'react'
+
+const InstructorPanel: React.FC = (): ReactNode => {
+    return (
+        <div>
+            <h1>Instructor Panel</h1>
+        </div>
+    )
+}
+
+export default InstructorPanel
